Move tests for filterWhileLoading
